// tss-api.ts
export interface components {
  schemas: {
    /** Empty Struct (Used mostly for Void RPC responses) */
    Unit: { [key: string]: unknown };
    /** Request batch update/create of placements, campaigns and bookings */
    BatchUpdateReq: {
      reqId?: string;
      placements?: components["schemas"]["ApiPlacement"][];
      campaigns?: components["schemas"]["ApiCampaign"][];
      bookings?: components["schemas"]["ApiBooking"][];
    };
    /** An advertisement placement */
    ApiPlacement: {
      identifier: components["schemas"]["ApiPlacementId"];
      name: components["schemas"]["strings_StringNE"];
      description: components["schemas"]["common_Notes"];
      summary?: string;
      searchContent?: string;
      category: components["schemas"]["db_PlacementCategory"];
      originalSpecs?: components["schemas"]["common_UrlResource"][];
      images?: components["schemas"]["common_UrlResource"][];
      documents?: components["schemas"]["common_UrlResource"][];
      location?: components["schemas"]["db_PlacementLocation"] | null;
      specifications: components["schemas"]["artworkspec_Specifications"];
      isDraft?: boolean;
      isVerified?: boolean;
      isDeprecated?: boolean;
      deprecatedAt?: components["schemas"]["Instant"] | null;
      activeStart?: components["schemas"]["LocalDate"] | null;
      activeEnd?: components["schemas"]["LocalDate"] | null;
      publicSearchable?: boolean;
      attributes?: components["schemas"]["artworkspec_Attributes"];
      version?: components["schemas"]["db_Version"];
      lastVerified?: components["schemas"]["Instant"] | null;
    };
// ...
    ApiCampaign: {
      campaignIdentifier: components["schemas"]["ApiCampaignId"];
      client: string;
      campaignName?: string;
    };
    ApiCampaignId: string;
    /** A placement booked at a date in a campaign */
    ApiBooking: {
      bookingIdentifier: components["schemas"]["ApiBookingId"];
      placementIdentifier: components["schemas"]["ApiPlacementId"];
      campaignIdentifier: components["schemas"]["ApiCampaignId"];
      liveDate: components["schemas"]["LocalDate"];
      endDate: components["schemas"]["LocalDate"];
    };
// ...
  };
}